tests: make neon_fixtures a bit thinner by splitting out some pageserver related helpers
